meliautoprofit/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.19
pytest==7.4.0
jinja2==3.1.2
python-telegram-bot==20.4
aiohttp==3.8.5
pytest-asyncio==0.21.1
tenacity==8.2.3
python-multipart==0.0.6
coverage==7.3.0
httpx==0.24.1
itsdangerous==2.1.2 